requirements.txt
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.4.0
Pillow>=8.3.0
opencv-python>=4.5.0
scikit-image>=0.18.0
jupyter>=1.0.0
notebook>=6.4.0
pandas>=1.3.0
seaborn>=0.11.0
